Rename MerkleNft -> MerkleUniqueToken, MerkleNftId -> MerkleUniqueTokenId
